only overwrite openvpn configuration when we have a openvpn tunnel config installed via gui (which has generated the openvpn.conf file). only then openvpn should be started. this avoids changing overlay md5sum after firmware upgrade

+++ files/common/usr/lib/ddmesh/ddmesh-init.d.sh
@@ -85,8 +85,11 @@ start() {
 	/usr/lib/ddmesh/ddmesh-privnet.sh start
 
 	logger -s -t $LOGGER_TAG "start service openvpn"
-	test -f /etc/config/openvpn.ffdd && mv /etc/config/openvpn.ffdd /etc/config/openvpn
-	test -x /etc/init.d/openvpn && /etc/init.d/openvpn start
+	# only touch any file when openvpn is used. else overlay md5sum will change
+	if [ -f "/etc/openvpn/openvpn.conf" ]; then
+		test -f /etc/config/openvpn.ffdd && mv /etc/config/openvpn.ffdd /etc/config/openvpn
+		test -x /etc/init.d/openvpn && /etc/init.d/openvpn start
+	fi
 
 	if [ -x /usr/bin/iperf3 ]; then
 		logger -s -t $LOGGER_TAG "start service iperf3"
